[metasrv] refactor: move meta api integration test to metasrv/tests/flight/metasrv_flight_meta_api.rs
